+++ include/ambition/Ambition_config.hh
@@ -67,11 +67,30 @@ public:
 	 const Mode mode
   );
 
+  /**
+   * The current running mode.
+   */
+  Mode currentMode() const;
+
 private:
   Mode _mode {Mode::Enhanced};
 };
 
 /** The global Ambition Config. */
 extern Config config;
+
+
+/**
+ * Detect whether the mode selection button was clicked and then run the mode
+ * selection screen.
+ *
+ * @return Whether anything was drawn to the screen and it needs to be redrawn.
+ */
+bool detectModeSelectClick();
+
+/**
+ * Draw the current running mode onto the top-right of the screen.
+ */
+void drawModeInformation();
 
 } // namespace Ambition